requirements.txt
aiogram==3.2.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
aiofiles==23.2.1